+++ test/declaration-block-single-line-max-declarations.js
@@ -22,6 +22,7 @@ testRule({
 			endColumn: 21,
 			rule,
 			url: undefined,
+			fix: undefined,
 			severity: `error`,
 			text: `Expected no more than 0 declarations (${rule})`,
 		},