+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 - The plugin now requires:
 	- `stylelint` version `17.0.0` or higher,
 	- `node.js` version `20.19.0` or higher.
+- Hex colors are no longer allowed.
 
 ### Added